Added script that checks links in a file.
All the links in src/data.js are hardcoded. This introduces the
possiblity of them breaking without anyone noticing. To be able
to automate the checking of all links this script is introduced.

The only parameter this script takes is a file path. The file is then
read and parsed for all occurences of http/https links. Those
links are then checked if they return an http status of 2** if not
the failing links are printed to the command line with their
returned status code.