Fixed request from PR review round 2
Disabled SSL verification for the `urlopen` call. This was necesarry,
since it created a false negativ, where an url was listed as broken
even though the server "just" used no current SSL version.

Rephrased parts of the scripts output to be more understandable for non
techy users (hopefully :)) and added the phrasing of the HTTP Errors to
the output.

Moved the except clause of the try statement (for opening the file)
closer together. The code in between was unnecessarly part of the try
statement because it had nothing to do with the opened file descriptor.