Basic CI definition for testing
Project now builds entirely in dockerfile
to add reproducability to the project.
To make this work the `.dockerignore` file
had to be deleted.

A `Cloudbuild.yaml` file has been added to
define building a docker image and deploying
in google cloud. The `CB_AUTH` secret used in
the build command is defined in the secret manager
of the google cloud.